@@ -4,11 +4,13 @@
     "license": "MIT",
     "description": "Fakturoid PHP SDK",
     "require": {
-        "php": "^7.1",
-        "guzzlehttp/guzzle": "6.3.*"
+        "php": "^7.4 || ^8.0",
+        "ext-json": "*",
+        "ext-fileinfo": "*",
+        "guzzlehttp/guzzle": "^7.4"
     },
     "require-dev": {
-        "phpunit/phpunit": "7.2.*"
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
