@@ -1,5 +1,6 @@
 <?php
 
+declare(strict_types=1);
 
 namespace Jahudka\FakturoidSDK\Endpoint;
 
@@ -8,28 +9,16 @@ use Jahudka\FakturoidSDK\Entity\Expense;
 
 
 /**
- * @method Expense get(int $id)
- * @method Expense[] getIterator(int $offset = null, int $limit = null)
- * @method Expense create(array $data)
- * @method Expense save(Expense $expense)
- * @method $this delete(Expense|int $expense)
- *
- * @method $this pay(Expense|int $expense, \DateTime|string|int $paidAt = null)
+ * @extends AbstractBillable<Expense>
+ * @method $this pay(Expense|int $expense, \DateTimeImmutable|string|int $paidAt = null)
  * @method $this removePayment(Expense|int $expense)
  */
 class Expenses extends AbstractBillable {
-
-    /**
-     * @param Client $api
-     */
     public function __construct(Client $api) {
         parent::__construct($api, 'accounts/' . $api->getSlug() . '/expenses', Expense::class);
     }
 
-    /**
-     * @return array
-     */
-    protected function getKnownOptions() {
+    protected function getKnownOptions(): array {
         return [
             'since',
             'updated_since',
@@ -42,10 +31,9 @@ class Expenses extends AbstractBillable {
     }
 
     /**
-     * @param string $variableSymbol
      * @return $this
      */
-    public function byVariableSymbol($variableSymbol) {
+    public function byVariableSymbol(string $variableSymbol) {
         return $this->setOption('variable_symbol', $variableSymbol);
     }
 }
