@@ -1,5 +1,6 @@
 <?php
 
+declare(strict_types=1);
 
 namespace Jahudka\FakturoidSDK\Endpoint;
 
@@ -9,28 +10,18 @@ use Jahudka\FakturoidSDK\Entity\Subject;
 
 
 /**
- * @method Subject get(int $id)
- * @method Subject[] getIterator(int $offset = null, int $limit = null)
- * @method Subject create(array $data)
- * @method Subject save(Subject $subject)
- * @method $this delete(Subject|int $subject)
+ * @extends AbstractEndpoint<Subject>
  */
 class Subjects extends AbstractEndpoint {
     use SearchableTrait,
         DateFilterableTrait,
         CustomFilterableTrait;
 
-    /**
-     * @param Client $api
-     */
     public function __construct(Client $api) {
         parent::__construct($api, 'accounts/' . $api->getSlug() . '/subjects', Subject::class);
     }
 
-    /**
-     * @return array
-     */
-    protected function getKnownOptions() {
+    protected function getKnownOptions(): array {
         return [
             'since',
             'updated_since',
@@ -38,5 +29,4 @@ class Subjects extends AbstractEndpoint {
             'query',
         ];
     }
-
 }
